Lift orbit action to zustand store
