+++ js/components/HomeScene/Scene/Rig/Actions.jsx
@@ -1,115 +1,22 @@
-import * as THREE from "three";
-import { useThree } from "@react-three/fiber";
 import { useControls, button } from "leva";
 
 import { useStore } from "@/js/lib/store";
-import App from "@/js/App";
-import { useEffect } from "react";
-
-const app = new App();
 
 export default function Actions() {
-    const getThreeState = useThree((state) => state.get);
-    const cameraPosition = new THREE.Vector3();
-    const boundingBox = new THREE.Box3();
+    const orbit = useStore((state) => state.orbit);
+    const resetCurrentBoxUuid = useStore((state) => state.resetCurrentBoxUuid);
 
     const resetCameraTarget = () => {
-        useStore.getState().resetCurrentBoxUuid();
-    };
-
-    /**
-     *
-     * @param {'right'|'up'|'down'|'left'} direction
-     */
-    const orbit = async (direction = "right") => {
-        const cameraControls = getThreeState().controls;
-        if (!cameraControls) return;
-
-        const scene = getThreeState().scene;
-        const padding = useStore.getState().padding;
-        const cameraTargetUuid = useStore.getState().cameraTargetUuid;
-        const cameraTarget = scene.getObjectByProperty(
-            "uuid",
-            cameraTargetUuid,
-        );
-        if (!cameraTarget.geometry) return;
-
-        // Store camera position to local Vector3
-        cameraControls.getPosition(cameraPosition, true);
-        // Calc bounds of camera target
-        cameraTarget.geometry.computeBoundingSphere();
-
-        // Store bounding box to local Box3
-        const boundingSphere = cameraTarget.geometry.boundingSphere.clone();
-        boundingSphere.getBoundingBox(boundingBox);
-
-        switch (direction) {
-            case "up": {
-                await cameraControls.rotate(
-                    0,
-                    THREE.MathUtils.degToRad(-90),
-                    true,
-                );
-                await cameraControls.fitToBox(cameraTarget, true, {
-                    paddingTop: padding,
-                    paddingRight: padding,
-                    paddingBottom: padding,
-                    paddingLeft: padding,
-                });
-                break;
-            }
-            case "down": {
-                await cameraControls.rotate(
-                    0,
-                    THREE.MathUtils.degToRad(90),
-                    true,
-                );
-                await cameraControls.fitToBox(cameraTarget, true, {
-                    paddingTop: padding,
-                    paddingRight: padding,
-                    paddingBottom: padding,
-                    paddingLeft: padding,
-                });
-                break;
-            }
-            case "left": {
-                await cameraControls.rotate(
-                    THREE.MathUtils.degToRad(-45),
-                    THREE.MathUtils.degToRad(cameraPosition.y > 0 ? 45 : -45),
-                    true,
-                );
-                break;
-            }
-            case "right":
-            default: {
-                await cameraControls.rotate(
-                    THREE.MathUtils.degToRad(45),
-                    THREE.MathUtils.degToRad(cameraPosition.y > 0 ? 45 : -45),
-                    true,
-                );
-                break;
-            }
-        }
+        resetCurrentBoxUuid();
     };
 
     useControls("Camera Actions", {
-        orbitRight: button(() => orbit("right")),
-        orbitLeft: button(() => orbit("left")),
-        orbitUp: button(() => orbit("up")),
-        orbitDown: button(() => orbit("down")),
+        orbitRight: button(async () => await orbit("right")),
+        orbitLeft: button(async () => await orbit("left")),
+        orbitUp: button(async () => await orbit("up")),
+        orbitDown: button(async () => await orbit("down")),
         resetCameraTarget: button(() => resetCameraTarget()),
     });
 
-    // bind event listeners
-    useEffect(() => {
-        const handleNavClick = (data) => {
-            const { direction } = data;
-            orbit(direction);
-        };
-        app.bus.on("component: HomeScene: nav: click", handleNavClick);
-        return () => {
-            app.bus.off("component: HomeScene: nav: click", handleNavClick);
-        };
-    }, []);
     return null;
 }
